Fix scorer to not treat musicians as instruments

+++ src/dto.rs
@@ -36,14 +36,14 @@ pub struct Placement {
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
 pub struct SolutionMetaDto {
     pub solver_name: String,
-    pub score: f64,
+    pub score: i64,
 }
 
 impl SolutionMetaDto {
     pub fn not_solved() -> Self {
         SolutionMetaDto {
             solver_name: "err_not_solved".to_string(),
-            score: 0.0,
+            score: 0,
         }
     }
 }
